requirements.txt
llama-index
llama-parse
nest_asyncio
streamlit
python-dotenv